Convert swagger.yaml to OAS3 (#25)
This is the second step of #23 after PR #24.

This PR passes the existing `swagger.yaml` (which is a Swagger 2.0 spec) through the Swagger 2.0 -> OpenAPI 3 conversion functionality of editor.swagger.io followed by a number of manual changes that reduce the unnecessary source line differences introduced by the conversion, such as reordering the properties of the objects in the schema. Minimizing the amount of diff makes it easier to review that the conversion didn't introduce any significant semantic change.

This PR also adds a GitHub action to validate the swagger.yaml on every commit.

Note: that editor.swagger.io doesn't report any issues with this swagger.yaml
Note 2: The `swagger.yaml` of this PR is still known to contain errors. In particular, it still doesn't reflect the changes from  [KIP 15](https://github.com/kadena-io/KIPs/blob/master/kip-0015.md), those will be addressed in later steps.

* Convert swagger.yaml to OAS3

* Reduce unnecessary git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Reduce git diff

* Define a GitHub action for validating the spec